requirements-core.txt
# Milton Overton AI Publicist - Core Dependencies
# Minimal dependencies for basic functionality
# Python 3.11+ required

# Core LLM & AI
anthropic>=0.34.0
openai>=1.3.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Database (SQLite for development)
aiosqlite>=0.19.0

# HTTP & Web Scraping
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0

# Security & Credentials
cryptography>=41.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Social Media (optional - install separately if needed)
# pip install linkedin-api tweepy instagrapi
